--- src/rcl/PanDialogInternals/Selection.styles.jsx
@@ -1,15 +1,0 @@
-export const sx = {
-  select: {
-      // "&.Mui-focused .MuiOutlinedInput-notchedOutline": {borderColor: "rgb(25, 118, 210)", borderWidth: "thin"},
-      background: 'white',
-  },
-  inputLabel: {
-    // background: 'purple',
-    // color: 'rgba(0, 0, 0, 0.6);',
-    // fontWeight: 400,
-    // fontSize: '0.875rem',
-    // "&.Mui-focused": {color: "rgb(25, 118, 210)"},
-  },
-}
-
-export default sx;
